Further changes to CI scripts to support SSL

+++ .ci_scripts/beforeInstall.sh
@@ -1,5 +1,7 @@
 #! /bin/bash
 
+set -e
+
 echo "[INFO] Clean some IVY cache"
 rm -rf "$HOME/.ivy2/local/org.reactivemongo"
 
@@ -30,21 +32,25 @@ fi
 
 echo "Max connection: $MAX_CON"
 
-# OpenSSL
-if [ ! -L "$HOME/ssl/lib/libssl.so.1.0.0" ]; then
-  cd /tmp
-  curl -s -o - https://www.openssl.org/source/openssl-1.0.1s.tar.gz | tar -xzf -
-  cd openssl-1.0.1s
-  rm -rf "$HOME/ssl" && mkdir "$HOME/ssl"
-  ./config -shared enable-ssl2 --prefix="$HOME/ssl" > /dev/null
-  make depend > /dev/null
-  make install > /dev/null
-else
-  rm -f "$HOME/ssl/lib/libssl.so.1.0.0" "libcrypto.so.1.0.0"
-fi
+if [ "$MONGO_PROFILE" = "self-ssl" -o "$MONGO_PROFILE" = "mutual-ssl" ]; then
 
-ln -s "$HOME/ssl/lib/libssl.so.1.0.0" "$HOME/ssl/lib/libssl.so.10"
-ln -s "$HOME/ssl/lib/libcrypto.so.1.0.0" "$HOME/ssl/lib/libcrypto.so.10"
+  # OpenSSL
+  if [ ! -L "$HOME/ssl/lib/libssl.so.1.0.0" ]; then
+    cd /tmp
+    curl -s -o - https://www.openssl.org/source/openssl-1.0.1s.tar.gz | tar -xzf -
+    cd openssl-1.0.1s
+    rm -rf "$HOME/ssl" && mkdir "$HOME/ssl"
+    ./config -shared enable-ssl2 --prefix="$HOME/ssl" > /dev/null
+    make depend > /dev/null
+    make install > /dev/null
+  else
+    rm -f "$HOME/ssl/lib/libssl.so.1.0.0" "libcrypto.so.1.0.0"
+  fi
+
+  ln -s "$HOME/ssl/lib/libssl.so.1.0.0" "$HOME/ssl/lib/libssl.so.10"
+  ln -s "$HOME/ssl/lib/libcrypto.so.1.0.0" "$HOME/ssl/lib/libcrypto.so.10"
+
+fi
 
 export LD_LIBRARY_PATH="$HOME/ssl/lib:$LD_LIBRARY_PATH"
 
@@ -84,14 +90,14 @@ else
     echo "  maxIncomingConnections: $MAX_CON" >> /tmp/mongod.conf
 fi
 
-mkdir /tmp/mongodb
+mkdir -p /tmp/mongodb
 
 SSL_PASS=""
 
 if [ "$MONGO_PROFILE" = "self-ssl" -o "$MONGO_PROFILE" = "mutual-ssl" ]; then
-    SSL_PASS=`uuidgen`
+    SSL_PASS="secret"
 
-    "$SCRIPT_DIR/genSslCert.sh"
+    "$SCRIPT_DIR/genSslCert.sh" $SSL_PASS
 
     cat >> /tmp/mongod.conf << EOF
   ssl:
@@ -110,7 +116,11 @@ EOF
     CAFile: $SCRIPT_DIR/client.pem
 EOF
 
-        # TODO: JKS https://blog.codecentric.de/en/2013/01/how-to-use-self-signed-pem-client-certificates-in-java/
+        keytool -importkeystore  -srcstoretype PKCS12 \
+            -srckeystore $SCRIPT_DIR/keystore.p12 \
+            -destkeystore /tmp/keystore.jks \
+            -storepass $SSL_PASS -srcstorepass $SSL_PASS
+
     fi
 fi
